requirements2.txt
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
distlib==0.3.6
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
filelock==3.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.4
redis==4.3.5
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.22.0
swapper==1.3.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
virtualenv==20.16.5
zope.interface==5.5.2
